Change graph.Graph and memory.Graph interfaces
Most of the interfaces from graph packages have been moved to memory
pacakge and are used to defined memory.Graph interface.

memory.Graph can now be passed as an option to memory.Store
