+++ pkg/graph/memory/memory.go
@@ -1,13 +1,56 @@
 package memory
 
-import "github.com/milosgajdos/netscrape/pkg/graph"
+import (
+	"context"
+
+	"github.com/milosgajdos/netscrape/pkg/graph"
+	"github.com/milosgajdos/netscrape/pkg/query"
+	"github.com/milosgajdos/netscrape/pkg/space"
+	"github.com/milosgajdos/netscrape/pkg/uuid"
+)
+
+// NOTE: this interface is a [temporary] hack!
+// Ideally, I would like to figure out how to parse
+// a generic GraphQL query into query.Query interface.
+// Or maybe graph.Query should simply accept a string,
+// which would then be parsed into query.Query.
+type Querier interface {
+	// Query the graph and return the results.
+	Query(context.Context, query.Query) ([]graph.Object, error)
+}
+
+// NodeAdder adds nodes to graph.
+type NodeAdder interface {
+	// NewNode returns a new node.
+	NewNode(context.Context, space.Entity, ...graph.Option) (graph.Node, error)
+	// AddNode adds a new node to graph.
+	AddNode(context.Context, graph.Node) error
+}
+
+// NodeRemover removes node from graph
+type NodeRemover interface {
+	// RemoveNode removes node from graph.
+	RemoveNode(context.Context, uuid.UID) error
+}
+
+// Linker links two nodes in graph.
+type Linker interface {
+	// Link links two nodes and returns the new edge.
+	Link(ctx context.Context, from, to uuid.UID, opts ...graph.Option) (graph.Edge, error)
+}
+
+// Unlinker removes link between two Nodes.
+type Unlinker interface {
+	// Unlink removes the link between the nodes with given UIDs from graph.
+	Unlink(ctx context.Context, from, to uuid.UID) error
+}
 
 // Graph is in-memory graph.
 type Graph interface {
 	graph.Graph
 	graph.SubGrapher
-	graph.NodeAdder
-	graph.NodeRemover
-	graph.Linker
-	graph.Unlinker
+	NodeAdder
+	NodeRemover
+	Linker
+	Unlinker
 }
