+++ pkg/store/store.go
@@ -23,8 +23,8 @@ type Entity interface {
 // Store must provide query capabilities by default.
 // Ideally, I would like to figure out how to parse
 // a generic GraphQL query into query.Query interface.
-// Or maybe store.Query should simply accept string,
-// which would then have to be parsed into query.Query.
+// Or maybe store.Query should simply accept a string,
+// which would then be parsed into query.Query.
 type Querier interface {
 	// Query store and return the results.
 	Query(context.Context, query.Query) ([]Entity, error)
@@ -36,10 +36,10 @@ type Store interface {
 	Graph(context.Context) (graph.Graph, error)
 	// Add Entity to store.
 	Add(context.Context, Entity, ...Option) error
-	// Link two entities in store.
-	Link(ctx context.Context, from, to uuid.UID, opts ...Option) error
 	// Delete Entity from store.
 	Delete(context.Context, Entity, ...Option) error
+	// Link two entities in store.
+	Link(ctx context.Context, from, to uuid.UID, opts ...Option) error
 	// Unlink two entities in store.
 	Unlink(ctx context.Context, from, to uuid.UID, opts ...Option) error
 }
